20260311/requirements.txt
# --- 核心基础与环境变量 ---
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.10.6

# --- LangChain生态 (用于工具Agent) ---
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10
langgraph==0.0.26  # (或你当前环境的具体版本，通常安装 langgraph 即可)
duckduckgo-search==8.1.1

# --- AutoGen生态 (0.10.x 新版异步架构) ---
autogen-agentchat==0.1.0  # 注：具体小版本号可能随你安装时略有不同
autogen-core==0.1.0
autogen-ext[openai]==0.1.0